\name{grid.density.limits}
\alias{grid.density.limits}
\title{Generates a density grid with a priori defined limits}
\usage{
  grid.density.limits(xvect, yvect, x.min, x.max, y.min,
    y.max)
}
\arguments{
  \item{xvect}{a numerical vector}

  \item{yvect}{a numerical vector}

  \item{x.min}{a minimum value for the x axis of the grid
  density object}

  \item{x.max}{a maximum value for the x axis of the grid
  density object}

  \item{y.min}{a minimum value for the y axis of the grid
  density object}

  \item{y.max}{a maximum value for the y axis of the grid
  density object}
}
\description{
  Generates a density grid with a priori defined limits
}

